Finalized authors tests refactor
